simulation/dependencies/remoter/pyproject.toml
[tool.poetry]
name = "remoter"
version = "0.17.2"
description = ""
authors = ["Simon Haendeler <simon.emanuel.haendeler@univie.ac.at>"]

[tool.poetry.dependencies]
python = ">=3.9"
trio = "^0.19.0"
pynng = "^0.7.1"
click = "^8.0.3"
tblib = "^1.7.0"
pyzmq = "^25.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[tool.poetry.group.dev.dependencies]
pytest-trio = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
